+++ relay-server/src/services/project_cache.rs
@@ -4,11 +4,12 @@ use std::sync::Arc;
 use std::time::Duration;
 
 use crate::extractors::RequestMeta;
-use crate::services::buffer::{EnvelopeBufferError, EnvelopeBufferGuard, GuardedEnvelopeBuffer};
+use crate::services::buffer::{EnvelopeBuffer, EnvelopeBufferError};
 use crate::services::processor::{
     EncodeMetrics, EnvelopeProcessor, MetricData, ProcessEnvelope, ProcessingGroup, ProjectMetrics,
 };
 use crate::services::project::state::UpstreamProjectState;
+use crate::Envelope;
 use chrono::{DateTime, Utc};
 use hashbrown::HashSet;
 use relay_base_schema::project::ProjectKey;
@@ -251,6 +252,9 @@ pub struct SpoolHealth;
 #[derive(Debug)]
 pub struct RefreshIndexCache(pub HashSet<QueueKey>);
 
+/// Handle an envelope that was popped from the envelope buffer.
+pub struct DequeuedEnvelope(pub Box<Envelope>);
+
 /// A cache for [`ProjectState`]s.
 ///
 /// The project maintains information about organizations, projects, and project keys along with
@@ -282,6 +286,7 @@ pub enum ProjectCache {
     UpdateSpoolIndex(UpdateSpoolIndex),
     SpoolHealth(Sender<bool>),
     RefreshIndexCache(RefreshIndexCache),
+    HandleDequeuedEnvelope(Box<Envelope>),
 }
 
 impl ProjectCache {
@@ -299,6 +304,7 @@ impl ProjectCache {
             Self::UpdateSpoolIndex(_) => "UpdateSpoolIndex",
             Self::SpoolHealth(_) => "SpoolHealth",
             Self::RefreshIndexCache(_) => "RefreshIndexCache",
+            Self::HandleDequeuedEnvelope(_) => "HandleDequeuedEnvelope",
         }
     }
 }
@@ -401,6 +407,15 @@ impl FromMessage<SpoolHealth> for ProjectCache {
 
     fn from_message(_message: SpoolHealth, sender: Sender<bool>) -> Self {
         Self::SpoolHealth(sender)
+    }
+}
+
+impl FromMessage<DequeuedEnvelope> for ProjectCache {
+    type Response = relay_system::NoResponse;
+
+    fn from_message(message: DequeuedEnvelope, _: ()) -> Self {
+        let DequeuedEnvelope(envelope) = message;
+        Self::HandleDequeuedEnvelope(envelope)
     }
 }
 
@@ -546,6 +561,7 @@ struct UpdateProjectState {
 /// Holds the addresses of all services required for [`ProjectCache`].
 #[derive(Debug, Clone)]
 pub struct Services {
+    pub envelope_buffer: Option<Addr<EnvelopeBuffer>>,
     pub aggregator: Addr<Aggregator>,
     pub envelope_processor: Addr<EnvelopeProcessor>,
     pub outcome_aggregator: Addr<TrackOutcome>,
@@ -553,30 +569,6 @@ pub struct Services {
     pub test_store: Addr<TestStore>,
     pub upstream_relay: Addr<UpstreamRelay>,
     pub global_config: Addr<GlobalConfigManager>,
-}
-
-impl Services {
-    /// Creates new [`Services`] context.
-    #[allow(clippy::too_many_arguments)]
-    pub fn new(
-        aggregator: Addr<Aggregator>,
-        envelope_processor: Addr<EnvelopeProcessor>,
-        outcome_aggregator: Addr<TrackOutcome>,
-        project_cache: Addr<ProjectCache>,
-        test_store: Addr<TestStore>,
-        upstream_relay: Addr<UpstreamRelay>,
-        global_config: Addr<GlobalConfigManager>,
-    ) -> Self {
-        Self {
-            aggregator,
-            envelope_processor,
-            outcome_aggregator,
-            project_cache,
-            test_store,
-            upstream_relay,
-            global_config,
-        }
-    }
 }
 
 /// Main broker of the [`ProjectCacheService`].
@@ -587,8 +579,6 @@ impl Services {
 struct ProjectCacheBroker {
     config: Arc<Config>,
     memory_checker: MemoryChecker,
-    // TODO: Make non-optional when spool_v1 is removed.
-    envelope_buffer: Option<Arc<GuardedEnvelopeBuffer>>,
     services: Services,
     // Need hashbrown because extract_if is not stable in std yet.
     projects: hashbrown::HashMap<ProjectKey, Project>,
@@ -743,10 +733,9 @@ impl ProjectCacheBroker {
         // Try to schedule unspool if it's not scheduled yet.
         self.schedule_unspool();
 
-        // TODO: write test that shows envelope can overtake when project becomes ready.
-        if let Some(buffer) = self.envelope_buffer.clone() {
-            tokio::spawn(async move { buffer.mark_ready(&project_key, true).await });
-        }
+        if let Some(envelope_buffer) = self.services.envelope_buffer.as_ref() {
+            envelope_buffer.send(EnvelopeBuffer::Ready(project_key))
+        };
     }
 
     fn handle_request_update(&mut self, message: RequestUpdate) {
@@ -785,12 +774,22 @@ impl ProjectCacheBroker {
     }
 
     fn handle_get(&mut self, message: GetProjectState, sender: ProjectSender) {
+        let GetProjectState {
+            project_key,
+            no_cache,
+        } = message;
         let project_cache = self.services.project_cache.clone();
-        self.get_or_create_project(message.project_key).get_state(
-            project_cache,
-            sender,
-            message.no_cache,
-        );
+        let envelope_buffer = self.services.envelope_buffer.clone();
+        let project = self.get_or_create_project(project_key);
+
+        // If the project is already loaded, inform the envelope buffer.
+        if !project.current_state().is_pending() {
+            if let Some(envelope_buffer) = envelope_buffer {
+                envelope_buffer.send(EnvelopeBuffer::Ready(project_key));
+            }
+        }
+
+        project.get_state(project_cache, sender, no_cache);
     }
 
     fn handle_get_cached(&mut self, message: GetCachedProjectState) -> ProjectState {
@@ -1080,21 +1079,19 @@ impl ProjectCacheBroker {
         }
     }
 
-    async fn peek_at_envelope(
+    fn handle_dequeued_envelope(
         &mut self,
-        mut peek: EnvelopeBufferGuard<'_>,
+        envelope: Box<Envelope>,
+        envelope_buffer: Addr<EnvelopeBuffer>,
     ) -> Result<(), EnvelopeBufferError> {
-        let envelope = peek.get().await?;
         if envelope.meta().start_time().elapsed() > self.config.spool_envelopes_max_age() {
-            let popped_envelope = peek.remove().await?;
             let mut managed_envelope = ManagedEnvelope::new(
-                popped_envelope,
+                envelope,
                 self.services.outcome_aggregator.clone(),
                 self.services.test_store.clone(),
                 ProcessingGroup::Ungrouped,
             );
             managed_envelope.reject(Outcome::Invalid(DiscardReason::Timestamp));
-            // TODO: metrics in all branches.
             return Ok(());
         }
         let sampling_key = envelope.sampling_key();
@@ -1106,14 +1103,10 @@ impl ProjectCacheBroker {
 
         // Check if project config is enabled.
         let project_info = match project_state {
-            ProjectState::Enabled(info) => {
-                peek.mark_ready(&own_key, true);
-                info
-            }
+            ProjectState::Enabled(info) => info,
             ProjectState::Disabled => {
-                let popped_envelope = peek.remove().await?;
                 let mut managed_envelope = ManagedEnvelope::new(
-                    popped_envelope,
+                    envelope,
                     self.services.outcome_aggregator.clone(),
                     self.services.test_store.clone(),
                     ProcessingGroup::Ungrouped,
@@ -1122,7 +1115,7 @@ impl ProjectCacheBroker {
                 return Ok(());
             }
             ProjectState::Pending => {
-                peek.mark_ready(&own_key, false);
+                envelope_buffer.send(EnvelopeBuffer::NotReady(own_key, envelope));
                 return Ok(());
             }
         };
@@ -1135,8 +1128,7 @@ impl ProjectCacheBroker {
                     .get_cached_state(services.project_cache, false),
             )
         }) {
-            Some((sampling_key, ProjectState::Enabled(info))) => {
-                peek.mark_ready(&sampling_key, true);
+            Some((_, ProjectState::Enabled(info))) => {
                 // Only set if it matches the organization ID. Otherwise treat as if there is
                 // no sampling project.
                 (info.organization_id == project_info.organization_id).then_some(info)
@@ -1146,7 +1138,7 @@ impl ProjectCacheBroker {
                 None
             }
             Some((sampling_key, ProjectState::Pending)) => {
-                peek.mark_ready(&sampling_key, false);
+                envelope_buffer.send(EnvelopeBuffer::NotReady(sampling_key, envelope));
                 return Ok(());
             }
             None => None,
@@ -1155,8 +1147,7 @@ impl ProjectCacheBroker {
         let project = self.get_or_create_project(own_key);
 
         // Reassign processing groups and proceed to processing.
-        let popped_envelope = peek.remove().await?;
-        for (group, envelope) in ProcessingGroup::split_envelope(*popped_envelope) {
+        for (group, envelope) in ProcessingGroup::split_envelope(*envelope) {
             let managed_envelope = ManagedEnvelope::new(
                 envelope,
                 services.outcome_aggregator.clone(),
@@ -1300,6 +1291,20 @@ impl ProjectCacheBroker {
                     ProjectCache::RefreshIndexCache(message) => {
                         self.handle_refresh_index_cache(message)
                     }
+                    ProjectCache::HandleDequeuedEnvelope(message) => {
+                        let envelope_buffer = self
+                            .services
+                            .envelope_buffer
+                            .clone()
+                            .expect("Called HandleDequeuedEnvelope without an envelope buffer");
+
+                        if let Err(e) = self.handle_dequeued_envelope(message, envelope_buffer) {
+                            relay_log::error!(
+                                error = &e as &dyn std::error::Error,
+                                "Failed to handle popped envelope"
+                            );
+                        }
+                    }
                 }
             }
         )
@@ -1311,7 +1316,6 @@ impl ProjectCacheBroker {
 pub struct ProjectCacheService {
     config: Arc<Config>,
     memory_checker: MemoryChecker,
-    envelope_buffer: Option<Arc<GuardedEnvelopeBuffer>>,
     services: Services,
     redis: Option<RedisPool>,
 }
@@ -1321,14 +1325,12 @@ impl ProjectCacheService {
     pub fn new(
         config: Arc<Config>,
         memory_checker: MemoryChecker,
-        envelope_buffer: Option<Arc<GuardedEnvelopeBuffer>>,
         services: Services,
         redis: Option<RedisPool>,
     ) -> Self {
         Self {
             config,
             memory_checker,
-            envelope_buffer,
             services,
             redis,
         }
@@ -1342,7 +1344,6 @@ impl Service for ProjectCacheService {
         let Self {
             config,
             memory_checker,
-            envelope_buffer,
             services,
             redis,
         } = self;
@@ -1352,7 +1353,6 @@ impl Service for ProjectCacheService {
 
         tokio::spawn(async move {
             let mut ticker = tokio::time::interval(config.cache_eviction_interval());
-            let mut report_ticker = tokio::time::interval(Duration::from_secs(1));
             relay_log::info!("project cache started");
 
             // Channel for async project state responses back into the project cache.
@@ -1423,7 +1423,6 @@ impl Service for ProjectCacheService {
             let mut broker = ProjectCacheBroker {
                 config: config.clone(),
                 memory_checker,
-                envelope_buffer: envelope_buffer.clone(),
                 projects: hashbrown::HashMap::new(),
                 garbage_disposal: GarbageDisposal::new(),
                 source: ProjectSource::start(
@@ -1479,32 +1478,12 @@ impl Service for ProjectCacheService {
                             broker.handle_message(message)
                         })
                     }
-                    peek = peek_buffer(&envelope_buffer) => {
-                        metric!(timer(RelayTimers::ProjectCacheTaskDuration), task = "peek_at_envelope", {
-                            if let Err(e) = broker.peek_at_envelope(peek).await {
-                                relay_log::error!(error = &e as &dyn std::error::Error, "Failed to peek envelope");
-                            }
-                        })
-                    }
-                    _ = report_ticker.tick() => {
-                        if let Some(envelope_buffer) = &envelope_buffer {
-                            metric!(gauge(RelayGauges::BufferPushInFlight) = envelope_buffer.inflight_push_count());
-                        }
-                    }
                     else => break,
                 }
             }
 
             relay_log::info!("project cache stopped");
         });
-    }
-}
-
-/// Temporary helper function while V1 spool exists.
-async fn peek_buffer(buffer: &Option<Arc<GuardedEnvelopeBuffer>>) -> EnvelopeBufferGuard {
-    match buffer {
-        Some(buffer) => buffer.peek().await,
-        None => std::future::pending().await,
     }
 }
 
@@ -1587,6 +1566,7 @@ mod tests {
         let (global_config, _) = mock_service("global_config", (), |&mut (), _| {});
 
         Services {
+            envelope_buffer: None,
             aggregator,
             envelope_processor,
             project_cache,
@@ -1616,8 +1596,6 @@ mod tests {
         .unwrap()
         .into();
         let memory_checker = MemoryChecker::new(MemoryStat::default(), config.clone());
-        let envelope_buffer =
-            GuardedEnvelopeBuffer::from_config(&config, memory_checker.clone()).map(Arc::new);
         let buffer_services = spooler::Services {
             outcome_aggregator: services.outcome_aggregator.clone(),
             project_cache: services.project_cache.clone(),
@@ -1640,7 +1618,6 @@ mod tests {
             ProjectCacheBroker {
                 config: config.clone(),
                 memory_checker,
-                envelope_buffer,
                 projects: hashbrown::HashMap::new(),
                 garbage_disposal: GarbageDisposal::new(),
                 source: ProjectSource::start(config, services.upstream_relay.clone(), None),