+++ relay-server/src/statsd.rs
@@ -26,10 +26,6 @@ pub enum RelayGauges {
     /// This metric is tagged with:
     /// - `reason`: Why keys are / are not unspooled.
     BufferPeriodicUnspool,
-    /// Number of envelopes currently waiting to be buffered.
-    ///
-    /// This corresponds to the number of corresponding tokio tasks currently scheduled or running.
-    BufferPushInFlight,
     /// The number of individual stacks in the priority queue.
     ///
     /// Per combination of `(own_key, sampling_key)`, a new stack is created.
@@ -60,7 +56,6 @@ impl GaugeMetric for RelayGauges {
             RelayGauges::BufferEnvelopesMemoryCount => "buffer.envelopes_mem_count",
             RelayGauges::BufferEnvelopesDiskCount => "buffer.envelopes_disk_count",
             RelayGauges::BufferPeriodicUnspool => "buffer.unspool.periodic",
-            RelayGauges::BufferPushInFlight => "buffer.push_inflight",
             RelayGauges::BufferStackCount => "buffer.stack_count",
             RelayGauges::BufferDiskUsed => "buffer.disk_used",
             RelayGauges::SystemMemoryUsed => "health.system_memory.used",
@@ -603,12 +598,11 @@ pub enum RelayCounters {
     ///  - `state_out`: The new state. `memory`, `memory_file_standby`, or `disk`.
     ///  - `reason`: Why a transition was made (or not made).
     BufferStateTransition,
-    /// Number of times the capacity is of the buffer is checked.
+    /// Number of envelopes that were returned to the envelope buffer by the project cache.
     ///
-    /// This metric is tagged with:
-    /// - `lock_acquired`: Whether the capacity check was done by acquiring the lock or using the
-    ///     old value.
-    BufferCapacityCheck,
+    /// This happens when the envelope buffer falsely assumes that the envelope's projects are loaded
+    /// in the cache and sends the envelope onward, even though the project cache cannot handle it.
+    BufferEnvelopesReturned,
     ///
     /// Number of outcomes and reasons for rejected Envelopes.
     ///
@@ -822,8 +816,8 @@ impl CounterMetric for RelayCounters {
             RelayCounters::BufferReadsDisk => "buffer.reads",
             RelayCounters::BufferEnvelopesWritten => "buffer.envelopes_written",
             RelayCounters::BufferEnvelopesRead => "buffer.envelopes_read",
+            RelayCounters::BufferEnvelopesReturned => "buffer.envelopes_returned",
             RelayCounters::BufferStateTransition => "buffer.state.transition",
-            RelayCounters::BufferCapacityCheck => "buffer.capacity_check",
             RelayCounters::Outcomes => "events.outcomes",
             RelayCounters::ProjectStateGet => "project_state.get",
             RelayCounters::ProjectStateRequest => "project_state.request",