+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Record too long discard reason for session replays. ([#3950](https://github.com/getsentry/relay/pull/3950))
 - Add `EnvelopeStore` trait and implement `DiskUsage` for tracking disk usage. ([#3925](https://github.com/getsentry/relay/pull/3925))
 - Increase replay recording limit to two hours. ([#3961](https://github.com/getsentry/relay/pull/3961))
+- Make EnvelopeBuffer a Service. ([#3965](https://github.com/getsentry/relay/pull/3965))
 
 **Internal**: